PieChart sort on key is broken
As the official demo shows it, the pie chart does not sort correctly its part by the key. "Q1" < "Q2" < "Q3" < "Q4" but not according to the chart.

The main reason is the call of the orderedGroup function, which assumes that the key value is provided when sorting a group, which is not the case.

Here is an attempt to correct it. A better version would include a way for the user to specify the sort function.