Don't draw ref-lines on top of y-axis
This only seems to affect chrome sometimes, for what it's worth, I've
never been able to reproduce this in safari or firefox.

The heart of the issue is that the ref-lines are drawn at the same
point as the y-axis, and the y-axis will disappear until the chart
is re-rendered.