Legend items show with dashed lines of the appropriate stroke-dasharray
This makes the most sense for a composite chart with several line charts that have different dash styles.

If any of the legend items have a stroke-dasharray property, all the legend items will show as lines.

Added an example with a composite chart that has different dash styles for each line graph.