changed data-table to support dynamically set columns for table header
Supplied 3 examples of setting the columns; the original method, which
requires the fixed html code for thead/cols; another which uses mostly
the data element names which become capitalized as col headings, and
another where most of the columns have the column label specified along
with an accessor function. There is some code to strip a
numberFormat/etc from a function converted to string, an attempt to
auto-generate a column header from a function, in data-table.js,
columnString().