major documentation cleanup
fixes #290: document all public API methods. the automatic verification is too noisy
to be sure, due to a possible bug in the script, and due the bazillion internal methods which are
not _prefixed.  but i think it's really all methods now.
fixes #615: handling line chart undefined/null values was implemented but not documented
fixes #610: disabling heatmap cell corner roundeed was implemented but not documented
fixes #441: document dc.filters

hundreds of grammatical fixes and clarifications too!