Fix vertical lines for ordinal charts
Charts using ordinal scales on the x axis with .renderVerticalGridLines
set did not work because ordinal scales do not have a .ticks function.
This patch checks for the existence of a .ticks function on the x scale
before trying to use it when rendering vertical lines, and if there isn't
one uses the scale's domain.