Correction on #766, fixes #909
running _computeOrderedGroups on the d3.stack data meant that the ordering
function had to have a different signature from usual, peering into the
inner `.data`.

this was causing a crossfilter crash in IE (#909).

so revert that and instead have `_ordinalXDomain` pluck the data and order
that. slightly more efficient for the other uses of `flattenStacks` that
don't need ordering, and actually correct.