[fix] Allow data inheritance in heatmaps
`selectAll` does not inherit data from its parent. This means that if the underlying crossfilter gets new data the selection
of `g.box-group` will see the new data, the children `title` and `rect` will not. The optical result of that is a seemingly emtpy
heatmap although the data has changed.