composite 2-pass drawing pre- & post- scale change
coord.plotData shall returns continuation if it wants to participate
the immediate action draws anything before the transition
the continuation will be run after the scale change

note there is nothing asynchronous here. instead we are using
continuations for the sake of capturing the joined selections that both
passes need.

2-pass is implemented but not specialized for the bar chart only

this commit also makes explicit that axis preparation and scale-setting
is different for render and redraw. render sets the scales twice,
probably to the same thing. this is because the scales must be valid
for the first pass, and they must *not* be set before a redraw, because
they must be consistent with the old view.

we'll see if this causes problems when clients (& maybe focus chart)
explicitly change the x domain. but if there are problems here, they are
probably not new. any enter selections on a too-soon scale change were
probably already screwed up.