remove callback from dc.transition
in preparation for #1116

this was a confusing parameter, invoked when it was decided that there
would be a transition. since we used it in exactly the case where we
were calling attrTween, and we already had to check that in another
place, seems clearer just to check that every time.

(it was also going to make it impossible to add an int/function-typed
parameter)