show scatter titles in these two examples
the scatter-series one i understand - we have to prevent the composite
chart from overriding the title function, since we need the special
title function that overrides the grotesque #702 scatter override of
key+value accessors

but i can't get why we don't need that for the multi-scatter
example. they're both composite charts, but this one works with the
shared composite title function because somehow or another, the key
accessor that title function sees is *not* the scatter override one.

it has something to do with the title function capturing key+value
accessors from its own closure, but i'm confused as heck and must be
tired as well.

ref #1200